kbuild: deb-pkg: Pass KBUILD_IMAGE explicitly to package make
KBUILD_IMAGE can be used to encourage the deb-pkg and bindeb-pkg rules
to package a non-standard kernel image. For example, for arm64 builds,
we can use KBUILD_IMAGE=arch/arm64/boot/Image to ensure the resulting
package includes an uncompressed version of the kernel.

However using this trick results in a orig, diff and dsc triplet that
does not rebuild the same package because MAKEFLAGS isn't set during
the rebuild. We can fix this by explicitly writing out KBUILD_IMAGE
to the intermediate files.

Signed-off-by: Daniel Thompson <daniel.thompson@linaro.org>